Major add-on of Algodrithmic Differentiation Types
Starting from the implementation of Daniel in Campiga, removed compile time variables and added vector as a basis for
the derivative, to allow for flexible computations at compile time. Extended the documentation to fit into the rest of
the library. Also extenranlised the defintions into an include file.

New unit tests, stand-alone quality is given in the file and the unit test.